# swagger.yaml
openapi: 3.0.3
info:
  title: Vidado - OpenAPI 3.0
  description: |-
    Está é a API destinada a aplicações que desejam se comunicar com a Vidado.
  contact:
    email: vidado.corp@yahoo.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: https://localhost:3000
...
paths:
...
  /signin/forgot:
    post:
      tags:
        - Usuário
      summary: Requisita código de recuperação
      description: Requisita código de recuperação
      operationId: createPasswordrecoveryCode
      requestBody:
        description: Requisita código de recuperação
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RequestRecoveryCode'
        required: true
      responses:
        '201':
          description: Código de recuperação enviado com sucesso
        '404':
          description: Email não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro de servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /signin/forgot/check-code:
    post:
      tags:
        - Usuário
      summary: Verifica código de recuperação
      description: Verifica código de recuperação
      operationId: checkRecoveryCode
      requestBody:
        description: Verifica código de recuperação
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CheckRecoveryCode'
        required: true
      responses:
        '202':
          description: Código aceito
        '401':
          description: Código inválido ou expirado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro de servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /signin/forgot/reset-password:
    patch:
      tags:
        - Usuário
      summary: Modifica a senha do usuário
      description: Modifica a senha do usuário
      operationId: resetPassword
      requestBody:
        description: Modifica a senha do usuário
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ResetPassword'
        required: true
      responses:
        '204':
          description: Senha modificada com sucesso
        '401':
          description: Código inválido ou expirado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro de servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /user/change-password:
    patch:
      tags:
        - Usuário
      summary: Requisita mudança de senha
      description: Requisita mudança de senha
      operationId: changePassword
      requestBody:
        description: Requisita mudança de senha
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ChangePassword'
        required: true
      responses:
        '202':
          description: Senha foi modificada com o sucesso
        '401':
          description: Senha atual incorreta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Email não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro de servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    RequestRecoveryCode:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          example: conta@email.com
    CheckRecoveryCode:
      type: object
      required:
        - email
        - code
      properties:
        email:
          type: string
          example: conta@email.com
        code:
          type: integer
          format: int32
          minLength: 4
          maxLength: 4
          minimum: 1000
          maximum: 9999
          example: 1234
    ResetPassword:
      type: object
      required:
        - email
        - code
        - new_password
        - confirm_password
      properties:
        email:
          type: string
          example: conta@email.com
        code:
          type: integer
          format: int32
          minLength: 4
          maxLength: 4
          minimum: 1000
          maximum: 9999
          example: 1234
        new_password:
          type: string
          minLength: 6
          maxLength: 255
          example: '123456'
        confirm_password:
          type: string
          minLength: 6
          maxLength: 255
          example: '123456'
    ChangePassword:
      type: object
      required:
        - email
        - password
        - new_password
        - confirm_password
      properties:
        email:
          type: string
          example: conta@email.com
        password:
          type: string
          minLength: 6
          maxLength: 255
          example: '123456'
        new_password:
          type: string
          minLength: 6
          maxLength: 255
          example: '123456'
        confirm_password:
          type: string
          minLength: 6
          maxLength: 255
          example: '123456'
...
    Error:
      type: object
      properties:
        errors:
          type: object
          properties:
            msg:
              type: string
              example: Mensagem de erro
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT